Refactor: 유틸리티 함수 분리 #101
